Made jest play nice with multiple platforms/import paths
